TMDM-3014: Optimization for code of new MDM Web UI
The optimization adds numeric codes to application wide MVC events. This allows us to use a switch statement to demux to the right method when we get an event instead of using a chain of if statements, which is currently being done now. BrowseRecordsEvents is changed to add codes. BrowseRecordsController and BrowseRecordsView are changed to demux by switching in their handleEvent() methods.

git-svn-id: http://talendforge.org/svn/tom/trunk@73344 f6f1c999-d317-4740-80b0-e6d1abc6f99e