TMDM-3014: Optimization for code of new MDM Web UI
Optimization that tries to reuse the ItemsDetailPanel (the one with gray outer borders and inner tab panel, but not the contents of the tab panel) if it already exists and is rendered. Currently, a new one is created everytime, including when the user is just switching from one record to another in the search results. By reusing the already rendered ItemsDetailPanel, we save a non-trivial amount of time from re-allocating and re-rendering it.

ItemsDetailPanel is changed so that clearing it when it is already rendered is a quick operation. Logic is added to BrowseRecordsView to reuse the ItemsDetailPanel when it is already rendered.

git-svn-id: http://talendforge.org/svn/tom/trunk@73362 f6f1c999-d317-4740-80b0-e6d1abc6f99e