TMDM-3014: Optimization for code of new MDM Web UI
Changed the display of tab labels in ItemsDetailPanel. Previously, used a very hackish method to measure the exact length of the tab label to truncate and add ellipsis if necessary. This was the only way to measure the exact width of rendered text unfortunately. Now we use a heuristic, where we simply use an estimate of the letter in the font (specifically we use the conservative estimate of the width of the letter "M" which is only narrower than the letter "W"). This has the advantage of being faster to calculate and also making it possible to run BrowseRecords by itself in debug mode. This has the disadvantage of not being as accurate, thus adding ellipsis sometimes when the tab may be wide enough to display all the text, or in the unlikely case when the label is "WWWWWWWWWW", truncating the label not early enough so that the ellipsis is not visible.

Also formatted code from previous commits using talend formatter. This is the only reason BrowseRecordsController and BrowseRecordsView are changed. ItemsDetailPanel is changed to use the new label width calculation method.

git-svn-id: http://talendforge.org/svn/tom/trunk@73372 f6f1c999-d317-4740-80b0-e6d1abc6f99e