TMDM-3014: Optimization for code of new MDM Web UI
In various methods, the same accessors are used over and over again to access the same thing. As I've read that access to non-local variables, using the dot notation, and function calls are expensive in JavaScript, I have replaced what I found with saving the result of the accessor in a local variable and reusing that local variable. Cumulatively this might have some impact.

git-svn-id: http://talendforge.org/svn/tom/trunk@73395 f6f1c999-d317-4740-80b0-e6d1abc6f99e