TMDM-3072 Safari and Chrome problems
This is a fix that makes sure BrowseRecords appears in Chrome and Safari. The JS code for BrowseRecords depends on the JS code for ItemBrowser. However, because the JS code for BrowseRecords is loaded asynchronously with the JS code for ItemBrowser, there is a change that it loads before the ItemBrowser JS, causing the error. The asynchronous load behavior for Safari and Chrome must be different enough from IE and FF to cause this error to appear frequently on those browsers.

This is a real error that is really located in mdm.webapp.general, which loads the separate JS modules like BrowseRecords bt does not take into consideration that one JS module might be loaded before another that it depends on. We have changed it to load ItemBrowser before BrowseRecords. We have also changed BrowseRecords to only use ItemBrowser after it has completely initialized itself, so that it gives more time for ItemBrowser to load and will still be initialized if ItemBrowser is not loaded. We have tested this on Safari and Chrome on WinXP and it always works.

mdm.webapp.general should be changed so that if a GWT JS module is loaded that depends on another JS module having been loaded, then the GWT module should set a timer that periodically checks whether the other module has been loaded, and if so, calls the initialization code. This is the fully technically correst solution, but it is complex and probably requires writing browser specific JSNI code.

git-svn-id: http://talendforge.org/svn/tom/trunk@74825 f6f1c999-d317-4740-80b0-e6d1abc6f99e